APOLLO-3946 Renamed all document types to be more descriptive as to directional flow
