+++ tests/OpgTest/Core/Model/Entity/Documents/OutgoingDocumentTest.php
@@ -1,16 +1,16 @@
 <?php
-namespace OpgTest\Core\Model\Entity\Correspondence;
+namespace OpgTest\Core\Model\Entity\Documents;
 
 use Opg\Core\Model\Entity\CaseItem\Lpa\Lpa;
 use Opg\Core\Model\Entity\CaseActor\Donor;
 use PHPUnit_Framework_TestCase;
-use Opg\Core\Model\Entity\Correspondence\Correspondence;
+use Opg\Core\Model\Entity\Documents\OutgoingDocument;
 use Opg\Common\Model\Entity\DateFormat as OPGDateFormat;
 
-class CorrespondenceTest extends PHPUnit_Framework_TestCase
+class OutgoingDocumentTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * @var \Opg\Core\Model\Entity\Correspondence\Correspondence;
+     * @var OutgoingDocument
      */
     private $correspondence;
 
@@ -28,7 +28,7 @@ class CorrespondenceTest extends PHPUnit_Framework_TestCase
 
     public function setUp()
     {
-        $this->correspondence = new Correspondence();
+        $this->correspondence = new OutgoingDocument();
     }
 
     public function testGetIterator()
